+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+- [#163](https://github.com/clojure-emacs/orchard/pull/163) Support Clojure Spec 2 in the Spec browser
+
 ## 0.10.0 (2022-09-04)
 
 ### Bugs fixed